Updates to use CE assistant API for all searches; Added/updates JSON-LD input classes; Import: Adding handling of pathways, transfer value;

+++ src/workIT.Factories/Entity_IdentifierValueManager.cs
@@ -21,13 +21,17 @@ namespace workIT.Factories
 	{
 		static string thisClassName = "Entity_IdentifierValueManager";
 		public static int CREDENTIAL_VersionIdentifier = 1;
-		public static int ORGANIZATION_AlternativeIdentifier = 2;
+		public static int ORGANIZATION_AlternativeIdentifier = 2; //OBSOLETE
 		public static int ASSESSMENT_VersionIdentifier = 3;
 		public static int LEARNING_OPP_VersionIdentifier = 4;
-		
+
+		public static int CREDENTIAL_Identifier = 11;
+		public static int ORGANIZATION_Identifier = 12;
+		public static int ASSESSMENT_Identifier = 13;
+		public static int LEARNING_OPP_Identifier = 14;
 
 		#region === Persistance ===================
-		public bool SaveList( List<ThisEntity> list, Guid parentUid, int IdentityValueTypeId, ref SaveStatus status )
+		public bool SaveList( List<ThisEntity> list, Guid parentUid, int IdentityValueTypeId, ref SaveStatus status, bool doingDelete )
 		{
             if ( !IsValidGuid( parentUid ) )
             {
@@ -42,7 +46,8 @@ namespace workIT.Factories
                 status.AddError( "Error - the parent entity was not found." );
                 return false;
             }
-            DeleteAll( parent, ref status );
+			if ( doingDelete )
+				DeleteAll( parent, ref status );
 
             if ( list == null || list.Count == 0 )
 				return true;
@@ -227,7 +232,7 @@ namespace workIT.Factories
 			to.EntityId = from.EntityId;
 			to.IdentityValueTypeId = from.IdentityValueTypeId;
 			to.Name = from.Name;
-			to.Description = from.Description;
+			//to.Description = from.Description;
 			to.IdentifierType = from.IdentifierType;
 			to.IdentifierValueCode = from.IdentifierValueCode;
 			to.Created = ( DateTime ) from.Created;
@@ -242,7 +247,7 @@ namespace workIT.Factories
 			
 			to.IdentityValueTypeId = from.IdentityValueTypeId;
 			to.Name = from.Name;
-			to.Description = from.Description;
+			//to.Description = from.Description;
 			to.IdentifierType = from.IdentifierType;
 			to.IdentifierValueCode = from.IdentifierValueCode;
